requirements.txt
numpy
pandas
geopandas
rasterio
shapely
torch
torchvision
open3d==0.13.0
matplotlib
scikit-learn
requests
tqdm
jupyterlab==3.6.8
notebook==6.4.12
setuptools>=60.2.0
pylas
rasterio
pillow
opencv-python
segment-anything
segmentation-models-pytorch
